python/adu-banding-teks/pyproject.toml
[tool.poetry]
name = "adu-banding-teks"
version = "0.1.0"
description = ""
authors = ["eka"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
google-cloud-speech = "^2.28.1"
ipykernel = "^6.29.5"
requests = "^2.32.3"
ffmpeg-python = "^0.2.0"
deepgram-sdk = "^3.7.6"
google-cloud-storage = "^2.18.2"
jiwer = "^3.0.5"
speechmatics-python = "^3.0.0"
whisper-normalizer = "^0.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
